--- core/src/main/java/com/flowci/core/credential/dao/SecretDao.java
@@ -1,33 +1,0 @@
-/*
- * Copyright 2018 flow.ci
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.flowci.core.credential.dao;
-
-import com.flowci.core.credential.domain.Secret;
-import org.springframework.data.mongodb.repository.MongoRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-/**
- * @author yang
- */
-@Repository
-public interface SecretDao extends MongoRepository<Secret, String>, CustomSecretDao {
-
-    Optional<Secret> findByName(String name);
-
-}
