+++ core/src/main/java/com/flowci/core/flow/service/FlowServiceImpl.java
@@ -21,8 +21,8 @@ import com.flowci.core.common.domain.Variables;
 import com.flowci.core.common.manager.SessionManager;
 import com.flowci.core.common.manager.SpringEventManager;
 import com.flowci.core.common.rabbit.RabbitChannelOperation;
-import com.flowci.core.credential.domain.Secret;
-import com.flowci.core.credential.service.SecretService;
+import com.flowci.core.secret.domain.Secret;
+import com.flowci.core.secret.service.SecretService;
 import com.flowci.core.flow.dao.FlowDao;
 import com.flowci.core.flow.dao.FlowUserDao;
 import com.flowci.core.flow.dao.YmlDao;
