+++ core/src/main/java/com/flowci/core/secret/domain/AuthSecret.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 flow.ci
+ * Copyright 2019 flow.ci
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,47 +14,40 @@
  * limitations under the License.
  */
 
-package com.flowci.core.credential.domain;
+package com.flowci.core.secret.domain;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.flowci.domain.SimpleKeyPair;
+import com.flowci.domain.SimpleAuthPair;
 import com.flowci.domain.SimpleSecret;
 import lombok.Getter;
 import lombok.Setter;
-import org.springframework.data.annotation.TypeAlias;
 import org.springframework.data.mongodb.core.mapping.Document;
 
 /**
+ * Username and password credential
+ *
  * @author yang
  */
 @Getter
 @Setter
 @Document(collection = "secret")
-public final class RSASecret extends Secret {
+public class AuthSecret extends Secret {
 
-    private SimpleKeyPair pair;
+    private SimpleAuthPair pair;
 
-    public RSASecret() {
-        this.pair = new SimpleKeyPair();
-        this.setCategory(Category.SSH_RSA);
+    public AuthSecret() {
+        this.pair = new SimpleAuthPair();
+        this.setCategory(Category.AUTH);
     }
 
     @JsonIgnore
-    public String getPrivateKey() {
-        return pair.getPrivateKey();
+    public String getUsername() {
+        return pair.getUsername();
     }
 
     @JsonIgnore
-    public String getPublicKey() {
-        return pair.getPublicKey();
-    }
-
-    public void setPrivateKey(String pk) {
-        this.pair.setPrivateKey(pk);
-    }
-
-    public void setPublicKey(String pk) {
-        this.pair.setPublicKey(pk);
+    public String getPassword() {
+        return pair.getPassword();
     }
 
     @Override
