+++ core/src/main/java/com/flowci/core/secret/domain/CreateRSA.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 flow.ci
+ * Copyright 2018 flow.ci
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,9 +14,10 @@
  * limitations under the License.
  */
 
-package com.flowci.core.credential.domain;
+package com.flowci.core.secret.domain;
 
-import com.flowci.domain.SimpleAuthPair;
+import com.flowci.domain.SimpleKeyPair;
+import com.google.common.base.Strings;
 import javax.validation.constraints.NotEmpty;
 import lombok.Data;
 
@@ -24,18 +25,23 @@ import lombok.Data;
  * @author yang
  */
 @Data
-public class CreateAuth {
+public class CreateRSA {
 
     @NotEmpty
     private String name;
 
     @NotEmpty
-    private String username;
+    private String publicKey;
 
     @NotEmpty
-    private String password;
+    private String privateKey;
 
-    public SimpleAuthPair getAuthPair() {
-        return SimpleAuthPair.of(username, password);
+    public boolean hasKeyPair() {
+        return !Strings.isNullOrEmpty(publicKey) && !Strings.isNullOrEmpty(privateKey);
     }
+
+    public SimpleKeyPair getKeyPair() {
+        return SimpleKeyPair.of(publicKey, privateKey);
+    }
+
 }
