+++ core/src/main/java/com/flowci/core/secret/service/SecretService.java
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-package com.flowci.core.credential.service;
+package com.flowci.core.secret.service;
 
-import com.flowci.core.credential.domain.AuthSecret;
-import com.flowci.core.credential.domain.Secret;
-import com.flowci.core.credential.domain.RSASecret;
+import com.flowci.core.secret.domain.AuthSecret;
+import com.flowci.core.secret.domain.Secret;
+import com.flowci.core.secret.domain.RSASecret;
 import com.flowci.domain.SimpleAuthPair;
 import com.flowci.domain.SimpleKeyPair;
