+++ core/src/test/java/com/flowci/core/test/flow/FlowServiceTest.java
@@ -19,9 +19,9 @@ package com.flowci.core.test.flow;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.flowci.core.common.domain.Variables;
-import com.flowci.core.credential.domain.AuthSecret;
-import com.flowci.core.credential.domain.RSASecret;
-import com.flowci.core.credential.service.SecretService;
+import com.flowci.core.secret.domain.AuthSecret;
+import com.flowci.core.secret.domain.RSASecret;
+import com.flowci.core.secret.service.SecretService;
 import com.flowci.core.flow.domain.Flow;
 import com.flowci.core.flow.domain.Flow.Status;
 import com.flowci.core.flow.domain.Yml;
